Fix: ForgetPassword css overwriting other forms
ForgetPassword.module.css was not scoping its form styles to the components and pages using it. Rather than targeting a class or the child of a class, the "form" tag was targeted. This caused it to override contact us form styles due to its higher level of specificity. To fix it I created a .form class in ForgetPassword.module.css and updated NewPassword and Reset forms to target the classname rather than the tag.